Reformatted Folders. Removed duplicates
There were a lot of folders that were duplicates, and files as well. I deleted a dozen or so. I did NOT delete all instances of a given material from the project though. So if you see an object that is untextured, the appropriate material does exist. Please either reapply the texture or let me know about it.